requirements.txt
chromadb
sentence-transformers
langchain
langchain-community
python-dotenv
pytest
hypothesis
openai
